surnia: Let's go thrifting
This adds a go config file on our device tree to avoid myself picking my commits to tweak it up on and on again.
So, what's different you may ask?

Theres a removal of ro.config.low_ram=true because people complains you about features like pip and notification access removed so it's a bad build because of that. Also tweaked the low ram lmk values to have a mid-normal mid-go value because there's people who treats a intel atom as it was a i5 and pretends to do high end gaming on a shitty device.
That's all